Add 1.0.0 mtl file backwards compatibility
In 1.0.0 the path in which 'OBJLoader' searched for MTL files was hardcoded
to `<assets root>/models/block`.
This was changed in 0263cb3 to just look in the models directory, but broke
compatibility with 1.0.0 MTL files.

This patch adds back that path as a fallback if `<assets root>/models/<model>.mtl`
could not be found initially